Suggest vignette edits
In "Installation", I would suggest providing a way to install the devel version of the package from the GitHub repository using devtools, and then complement with the Bioconductor instructions once the package is submitted.
In "Method summary", I would suggest adding a schematic figure where the top window vs decreasing expression windows correlation system is explained in an intuitive, visual way.
In "Correlation of the bins" I would indicate the type of correlation that we are implementing, which is Pearson, if I'm correct? I wasn't sure since you might have changed something, so I didn't add it myself.
In addition, I have edited the text for clarity and removed a few type-os.